Extract implementation of OAuth2 provider
This is preparation for creation of a new Open ID Connect provider,
which will now be able to reuse most of the logic for the OAuth2
provider.
